Revert "Add `sdk/rust/oak_runtime_compat` (#405)"
This reverts commit 34e738b503d9720c6e3ec45e8ccc8435419b2006.
