--- sdk/rust/oak_runtime_compat/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "oak_runtime_compat"
-version = "0.1.0"
-authors = ["Ben Blaxill <benblaxill@google.com>"]
-edition = "2018"
-
-[lib]
-name = "oak_runtime_compat"
-
-[dependencies]
-libc = "*"
-
-[workspace]
